Remove `vm` module in favor of flat file
